+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Master
 
+- Display an error message when polling fails on bidding - yuki24
 - Fixes an issue where placing a bid hangs - yuki24
 
 ### 1.12.9